+++ src/renderer/dto/render-command.dto.ts
@@ -12,6 +12,9 @@ export class RenderCommandDto {
     @ApiProperty({ description: "Whether to forward request headers" })
     forwardHeaders?: boolean = false
 
+    @ApiProperty({ description: "Whether to spoof user-agent" })
+    spoofUserAgent?: boolean = false
+
     @ApiProperty({ description: "Whether to use rotating HTTP proxies" })
     httpProxy?: boolean = false
 }