+++ src/renderer/dto/render-query.dto.ts
@@ -5,5 +5,6 @@ export class RenderQueryDto extends PickType(RenderCommandDto, [
     "url",
     "blockAds",
     "forwardHeaders",
+    "spoofUserAgent",
     "httpProxy",
 ]) {}