+++ src/resources/types/user-agent.interface.ts
@@ -1,0 +1,5 @@
+export interface UserAgentEntity {
+    platform: string
+    deviceCategory: string
+    userAgent: string
+}